+++ src/reaction-role/constants.js
@@ -23,13 +23,20 @@ const ReactionRoleEvent = Object.freeze({
 /**
  * Requirement type to win some role.
  * @typedef {Object} RequirementType
- * @property {string} BOOST - Need be a booster to win this role.
- * @property {string} VERIFIED_DEVELOPER - Need be a verified developer to win this role.
+ * @property {number} [BOOST=1] - Need be a booster to win this role.
+ * @property {number} [VERIFIED_DEVELOPER=2] - Need be a verified developer to win this role.
+ * @property {number} [PERMISSION=3] - Need has some permissions to win this role.
+ * @property {number} [ROLES=4] - Need has all allow listed roles and hasn't all denied listed roles.
+ * @property {number} [USERS=5] - Need be inluded in allow list and not included in deny list to win this role.
  * @readonly
  */
 const RequirementType = Object.freeze({
-    BOOST: 'BOOST',
-    VERIFIED_DEVELOPER: 'VERIFIED_DEVELOPER',
+    UNKNOWN: 0,
+    BOOST: 1,
+    VERIFIED_DEVELOPER: 2,
+    PERMISSION: 3,
+    ROLES: 4,
+    USERS: 5,
 });
 
 /**
@@ -66,11 +73,11 @@ const ActionType = Object.freeze({
 });
 
 /**
- * Check if is a valid reaction role type.
- * @param {ReactionRoleType | number} type - Reaction role type to check.
+ * Check if a number is valid reaction role type.
+ * @param {ReactionRoleType} number - Type of reaction role to check if it's valid.
  * @return {boolean}
  */
-const isValidReactionRoleType = (type) => !isNaN(type) && (type >= ReactionRoleType.NORMAL && type <= ReactionRoleType.REVERSED);
+const isValidReactionRoleType = (number) => !isNaN(number) && (number >= ReactionRoleType.NORMAL && number <= ReactionRoleType.REVERSED);
 
 module.exports = {
     RequirementType,